Added functionalities:
refresh the queues if an event is changed after racers have already been
entered
PRINT command now makes printer print the race results. Printer class
now has a boolean flag called 'log' where it won't print console crap
unless the command has been issued to Chronotimer. A button or something
needs to be added to the GUI to allow for that. Additionally, a PRINT
command needs to be added to the function screen.